Test moved to test_triangulation.py
